find in excel and create a new file with it
This code was created to filter an excel file for specific order IDs and create a new file containing only the matching rows. It was designed to be used in a business context where certain orders needed to be ticketed for further action. If there are no matching rows, the code will create a text file to indicate that no business needs to be ticketed. This code can be easily customized to work with other search criteria and file paths